update save script to optionally use sudo

+++ copied/save.sh
@@ -5,7 +5,7 @@ set -euo pipefail
 SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
 pushd "$SCRIPT_DIR" > /dev/null
 
-echo "[+] Copying files..."
+echo "[+] Copying files (this might need root access)..."
 
 for file in $(cat files.txt); do
   # make containing folder structure
@@ -13,8 +13,12 @@ for file in $(cat files.txt); do
   parent_dir="${parent_dir#/}" # remove leading slash of dir
   mkdir -p "$parent_dir"
 
-  # copy the file
-  sudo cp $file $parent_dir
+  # copy each file, only use root if necessary
+  if [ -f $file ]; then
+    cp $file $parent_dir
+  else
+    sudo cp $file $parent_dir
+  fi
 done
 
 popd > /dev/null
